[MOVE/#25] type폴더로 변경
